backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.35
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# File parsing
pdfplumber>=0.11.4
openpyxl>=3.1.5
python-docx>=1.1.2

# Embeddings & Vector search
faiss-cpu>=1.8.0
sentence-transformers>=3.1.1
numpy>=1.26.4

# LLM (optional – requires OPENAI_API_KEY)
openai>=1.51.0

# Export
reportlab>=4.2.2

# HTTP (for optional Sanity integration)
httpx>=0.27.2

# Env
python-dotenv>=1.0.1

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
